Fix rerolls, especially those involving chaos.
Set the flags correctly so that the game knows we're rerolling
on the boon menu if that's the case. Also, break once the
first line is selected in this case, because it's threaded, so
the rewards will be rerolled before it finishes playing and a
subsequent one (if any) needs to be selected.