Update to v1.37332, and finish fixing rerolls.
A new chance in the 1.37332 patch clears the rolled boons in combat rooms
after the item drops and causes them to be rolled when you pick up the
reward instead. When this happens, the RNG is reset to the number of
rewards of the same type you've picked up in the current run (including
this one).

To finish fixing rerolls, pass in the excluded item name instead of the
excluded item object to SetTraitsOnLoot.