Fix bug with Elysium Midshop generation when the previous room has a challenge switch.
IsGameStateEligible doesn't always use the passed-in currentRun for its calculations;
sometimes (due to a typo in the official game code?) it uses the CurrentRun global
instead. These are different when running predictions. In this case, the check for
whether Flame Wheel Release was eligible for shop generation used CurrentRun.CurrentRoom
and therefore mistakenly thought that the challenge switch was in the shop; this in
turn cause the prediction mod not to roll to see if Flame Wheel Release would appear.

Fix by wrapping IsGameStateEligible globally to set CurrentRun to currentRun at the
start of the function and reset it to its previous value at the end.